telegram_bot/requirements.txt
aiogram>=3.0.0
apscheduler
aiosqlite
httpx
python-dotenv
pytz
